backend/pyproject.toml
[project]
name = "univy"
version = "0.1.0"
description = ""
authors = [
    {name = "kunkoala",email = "azharffrahadian@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi[standard] (>=0.115.9,<0.116.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "redis (>=6.0.0,<7.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "aioredis (>=2.0.1,<3.0.0)",
    "celery>=5.5.2,<6.0.0",
    "docling>=2.38.1",
    "loguru>=0.7.3",
    "lightrag-hku>=1.3.9",
    "openai>=1.93.0",
    "graspologic>=3.4.1",
    "nano-vectordb>=0.0.4.3",
]

[tool.poetry]
packages = [{include = "univy", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
